+++ app/scripts/TiledPixiTrack.js
@@ -238,6 +238,12 @@ class TiledPixiTrack extends PixiTrack {
 
       if (isTilesetInfo(tilesetInfo)) {
         this.tilesetInfo = tilesetInfo;
+
+        if (this.tilesetInfo.chromsizes) {
+          // We got chromosome info from the tileset info so let's parse it
+          // into an object we can use
+          this.chromInfo = parseChromsizesRows(this.tilesetInfo.chromsizes);
+        }
       } else {
         // no tileset info for this track
         console.warn(
@@ -283,6 +289,10 @@ class TiledPixiTrack extends PixiTrack {
 
       this.refreshTiles();
 
+      // Let this track know that tileset info was received
+      this.tilesetInfoReceived();
+
+      // Let external listeners know that tileset info was received
       if (handleTilesetInfoReceived) handleTilesetInfoReceived(tilesetInfo);
 
       // @ts-expect-error This should never happen since options is set in Track
@@ -397,6 +407,13 @@ class TiledPixiTrack extends PixiTrack {
       }
     }
   }
+
+  /** Called when tileset info is received. The actual tileset info
+   * can be found in this.tilesetInfo.
+   *
+   * Child tracks can implement this method.
+   */
+  tilesetInfoReceived() {}
 
   /**
    * Return the set of ids of all tiles which are both visible and fetched.